Fetch new data and store it in the database (untested)
